requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=10.0
httpx>=0.25.0
pyyaml>=6.0
pydantic>=2.5.0

# Streaming support
sse-starlette>=1.8.0

# Monitoring
psutil>=5.9.0
prometheus-client>=0.18.0

# ML Backend (optional - uncomment as needed)
# torch>=2.0.0
# transformers>=4.35.0
# vllm>=0.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# WebSocket support
python-multipart>=0.0.6
